+++ ceammc/CHANGELOG.md
@@ -10,7 +10,7 @@
   - conv.char2morse - single character to morse code converter
   - proto.morse - morse encoder with specified speed and other settings
 - new object inlets:
-  - metro.seq - second inlet added to set sequence pattern
+  - metro.seq - second inlet added to set interval and third inlet to set pattern
 - new object methods:
   - euclid method added to ui.radio to set euclid pattern when in checklist mode
   - hexbeat method added to ui.radio to set hexbeat pattern when in checklist mode